Undo 29_DummyDefault en afvangen van backup zip file export DwCA

+++ nl.naturalis.nda.export/src/main/java/nl/naturalis/nda/export/dwca/ExportDwCAUtilities.java
@@ -176,12 +176,15 @@ public class ExportDwCAUtilities {
 		InputStream inputstream = null;
 		OutputStream outputstream = null;
 		try {
-			inputstream = new FileInputStream(sourceFile);
-			outputstream = new FileOutputStream(DestinationFile);
-			byte[] buffer = new byte[2048];
-			int length;
-			while ((length = inputstream.read(buffer)) > 0) {
-				outputstream.write(buffer, 0, length);
+			if (sourceFile.exists())
+			{	
+				inputstream = new FileInputStream(sourceFile);
+				outputstream = new FileOutputStream(DestinationFile);
+				byte[] buffer = new byte[2048];
+				int length;
+				while ((length = inputstream.read(buffer)) > 0) {
+					outputstream.write(buffer, 0, length);
+				}
 			}
 		}
 		finally {
